No more js file
