+++ src/screens/MapPage.tsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import {StyleSheet, View, StatusBar} from 'react-native';
+import MapView, {PROVIDER_GOOGLE} from 'react-native-maps';
+
+export default function MapPage() {
+  return (
+    <View style={styles.container}>
+      <StatusBar barStyle={'dark-content'} />
+      <MapView
+        provider={PROVIDER_GOOGLE}
+        style={styles.map}
+        region={{
+          latitude: 11.93623,
+          longitude: 108.445259,
+          latitudeDelta: 0.01,
+          longitudeDelta: 0.01,
+        }}
+      />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  map: {
+    ...StyleSheet.absoluteFillObject,
+  },
+});
